Optimize stakeNFT storage (#24)
* Optimize stakeNFT storage

* Remove unnecessary comment

* Update storage to optimize the collect functions

+++ src/StakeNFT.sol
@@ -33,7 +33,7 @@ contract StakeNFT is ERC721, MagicValue, Admin, Governance, CircuitBreaker, Atom
         uint32 freeAfter;
 
         // block number after which the position may be collected or burned.
-        uint32 withdrawFreeAfter;
+        uint256 withdrawFreeAfter;
 
         // the last value of the ethState accumulator this account performed a
         // withdraw at
@@ -334,7 +334,7 @@ contract StakeNFT is ERC721, MagicValue, Admin, Governance, CircuitBreaker, Atom
     function _lockWithdraw(uint256 tokenID_, uint256 duration_) internal returns(uint256 shares) {
         require(_exists(tokenID_), "StakeNFT: Token ID doesn't exist!");
         Position memory p = _positions[tokenID_];
-        uint32 freeDur = uint32(block.number) + uint32(duration_);
+        uint256 freeDur = block.number + duration_;
         p.withdrawFreeAfter = freeDur > p.withdrawFreeAfter ? freeDur : p.withdrawFreeAfter;
         _positions[tokenID_] = p;
         return p.shares;
